pyproject.toml
[project]
name = "health-coach"
version = "0.1.0"
description = "Persönlicher Telegram Gesundheitscoach powered by Claude"
requires-python = ">=3.12,<3.14"
dependencies = [
    "python-telegram-bot==21.6",
    "anthropic>=0.49.0",
    "apscheduler==3.10.4",
    "pyyaml==6.0.2",
    "python-dotenv==1.0.1",
]

